+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2026-02-18
+### Added
+- Enabling of some required Google APIs
+- Variables `private_network_name` and `private_subnet_name` (both default: `default`) which you can set to create private connected to a non-default VPC
+
 ## [0.6.0] - 2025-12-06
 ### Added
 - `examples/mcp-toolbox`